requirements.txt
# bot_new/requirements.txt
aiogram>=3.7.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
# requests удален, так как не используется